Changed variable names for New Game button and Status Label to be more cohesive to their utility.
Original logic for handling the New Game button was moved to a resetGame() function and refined using a for() loop.

Original logic for checking a winner was a gigantic list of nested conditional statements. That has been refined down into a single conditional that compares the indices in the b[] array to multidimensional array of all possible wins.

Added functionality to check for a draw if all buttons have been actioned (filled) with no win condition found.