Converted to using javax.swing alongside java.awt.
Originally, there was an issue in creating the victory line across the winning combo. Using javax.swing allows the creation of a layered UI. The victory line and it's logic is handled within an OverlayPanel that sits above the board game panel, which houses the buttons.

Also changed the commentation and organization of the panels, game logic, and main function to be more properly formatted and easier to read.